requirements.txt
asgiref==3.8.1
certifi==2024.7.4
charset-normalizer==3.3.2
clocker==0.1
Django==5.0.7
django-ckeditor5==0.0.4
django-debug-toolbar==4.4.6
django-js-asset==2.2.0
django-multiselectfield==0.1.13
et-xmlfile==1.1.0
idna==3.7
numpy==2.0.1
openpyxl==3.1.5
pandas==2.2.2
pillow==10.4.0
psycopg2-binary==2.9.9
python-dateutil==2.9.0.post0
pytz==2024.1
requests==2.32.3
setuptools==72.2.0
six==1.16.0
sqlparse==0.5.1
tzdata==2024.1
urllib3==2.2.2
